requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.46
python-dotenv==0.15.0
Werkzeug==1.0.1
# for heroku
gunicorn==20.0.4